+++ src/components/popover/popover.ts
@@ -117,7 +117,7 @@ export class QuietPopover extends QuietElement {
       }
 
       // Setup the new anchor
-      this.anchor = this.for ? root.querySelector(`#${this.for}`) : null;
+      this.anchor = this.for ? root.getElementById(this.for) : null;
 
       if (this.anchor) {
         this.anchor.addEventListener('click', this.handleAnchorClick);